+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file.
   * Duplicating Aggregators.
   * Editing the title and description of view in the panel.
 * Pressing 'Close' button in form constructors will open previous page.
+* Fix autoresize of uml package object.
 
 ## [0.2.0] - 2018-06-26
 ### Added